requirements.txt
openai>=1.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.11.6
tiktoken>=0.5.0
python-dotenv>=1.0.0
lxml>=4.9.0
pytest>=7.4.0

